fix: delete the old rt trip updates, and rename the new stuff
