--- airflow/dags/rt_loader/parse_rt_trip_updates_json.yml
@@ -1,8 +1,0 @@
-operator: operators.RealtimeToFlattenedJSONOperator
-provide_context: true
-
-filename_prefix: tu
-rt_file_substring: trip_updates
-
-dependencies:
-  - external_trip_updates
