+++ airflow/plugins/operators/realtime_to_flattened_json_operator.py
